@@ -1,5 +1,11 @@
 module github.com/lxzan/memorycache
 
-go 1.16
+go 1.18
 
 require github.com/stretchr/testify v1.8.1
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
