@@ -1,108 +1,149 @@
 package memorycache
 
 import (
-	"github.com/lxzan/memorycache/internal/hashmap"
 	"github.com/lxzan/memorycache/internal/heap"
 	"github.com/lxzan/memorycache/internal/types"
 	"github.com/lxzan/memorycache/internal/utils"
+	"math"
+	"strings"
+	"sync"
 	"time"
 )
 
 type MemoryCache struct {
-	config  *Config
-	storage *hashmap.ConcurrentMap
+	config  *types.Config
+	storage []*bucket
 }
 
+// New 创建缓存数据库
 func New(options ...Option) *MemoryCache {
-	var config = &Config{}
+	var config = &types.Config{}
 	options = append(options, withInitialize())
 	for _, fn := range options {
 		fn(config)
 	}
 
-	return &MemoryCache{
+	mc := &MemoryCache{
 		config:  config,
-		storage: hashmap.NewConcurrentMap(config.Segment, config.TTLCheckInterval),
+		storage: make([]*bucket, config.BucketNum),
 	}
+
+	for i, _ := range mc.storage {
+		mc.storage[i] = &bucket{
+			Map:  make(map[string]*types.Element, config.InitialSize),
+			Heap: heap.New(config.InitialSize),
+		}
+	}
+
+	go func() {
+		var ticker = time.NewTicker(config.Interval)
+		defer ticker.Stop()
+		for {
+			<-ticker.C
+			for _, bucket := range mc.storage {
+				bucket.expireTimeCheck(config.MaxKeysDeleted, config.MaxCapacity)
+			}
+		}
+	}()
+
+	return mc
 }
 
-func (c *MemoryCache) valid(now, t int64) bool {
-	return t <= 0 || t > now
+func (c *MemoryCache) getBucket(key string) *bucket {
+	var idx = utils.NewFnv32(key) & (c.config.BucketNum - 1)
+	return c.storage[idx]
 }
 
-func (c *MemoryCache) getExpireTimestamp(expiration time.Duration) int64 {
-	return time.Now().Add(expiration).UnixNano() / 1000000
+// 获取过期时间, d<=0表示永不过期
+func (c *MemoryCache) getExp(d time.Duration) int64 {
+	if d <= 0 {
+		return math.MaxInt
+	}
+	return time.Now().Add(d).UnixMilli()
 }
 
-// 设置键值和过期时间
-// expiration: <=0表示永不过期
-func (c *MemoryCache) Set(key string, value interface{}, expiration time.Duration) {
-	var ele = types.Element{
+// 设置键值和过期时间. exp<=0表示永不过期.
+func (c *MemoryCache) Set(key string, value any, exp time.Duration) (replaced bool) {
+	var b = c.getBucket(key)
+	b.Lock()
+	defer b.Unlock()
+
+	var expireAt = c.getExp(exp)
+	v, ok := b.Map[key]
+	if ok {
+		v.Value = value
+		v.ExpireAt = expireAt
+		b.Heap.Down(v.Index, b.Heap.Len())
+		return true
+	}
+
+	var ele = &types.Element{
+		Key:      key,
 		Value:    value,
-		ExpireAt: c.getExpireTimestamp(expiration),
+		ExpireAt: expireAt,
 	}
 
-	var bucket = c.storage.GetBucket(key)
-	bucket.Lock()
-	bucket.Map[key] = ele
-	if ele.ExpireAt != -1 {
-		bucket.Heap.Push(heap.Element{
-			Key:      key,
-			ExpireAt: ele.ExpireAt,
-		})
-	}
-	bucket.Unlock()
+	b.Heap.Push(ele)
+	b.Map[key] = ele
+	return false
 }
 
-// 根据键获取值
-func (c *MemoryCache) Get(key string) (interface{}, bool) {
-	var bucket = c.storage.GetBucket(key)
-	bucket.RLock()
-	result, exist := bucket.Map[key]
-	bucket.RUnlock()
-	if !exist || !c.valid(utils.Timestamp(), result.ExpireAt) {
+// Get 获取
+func (c *MemoryCache) Get(key string) (any, bool) {
+	var b = c.getBucket(key)
+	b.Lock()
+	v, exist := b.Map[key]
+	b.Unlock()
+	if !exist || v.Expired(time.Now().UnixMilli()) {
 		return nil, false
 	}
-	return result.Value, true
+	return v.Value, true
+}
+
+// GetAndRefresh 获取. 如果存在, 刷新过期时间.
+func (c *MemoryCache) GetAndRefresh(key string, exp time.Duration) (any, bool) {
+	var b = c.getBucket(key)
+	b.Lock()
+	defer b.Unlock()
+
+	v, exist := b.Map[key]
+	if !exist || v.Expired(time.Now().UnixMilli()) {
+		return nil, false
+	}
+
+	v.ExpireAt = c.getExp(exp)
+	b.Heap.Down(v.Index, b.Heap.Len())
+	return v, true
 }
 
 // 删除一个键
-func (c *MemoryCache) Delete(key string) {
-	var bucket = c.storage.GetBucket(key)
-	bucket.Lock()
-	delete(bucket.Map, key)
-	bucket.Unlock()
-}
+func (c *MemoryCache) Delete(key string) (deleted bool) {
+	var b = c.getBucket(key)
+	b.Lock()
+	defer b.Unlock()
 
-// 设置键的过期时间
-func (c *MemoryCache) Expire(key string, expiration time.Duration) {
-	var bucket = c.storage.GetBucket(key)
-	bucket.Lock()
-	if result, exist := bucket.Map[key]; exist && c.valid(utils.Timestamp(), result.ExpireAt) {
-		result.ExpireAt = c.getExpireTimestamp(expiration)
-		bucket.Map[key] = result
-		if result.ExpireAt > 0 {
-			bucket.Heap.Push(heap.Element{
-				Key:      key,
-				ExpireAt: result.ExpireAt,
-			})
-		}
+	v, ok := b.Map[key]
+	if !ok {
+		return false
 	}
-	bucket.Unlock()
+
+	b.Heap.Delete(v.Index)
+	delete(b.Map, key)
+	return true
 }
 
-// 获取所有键
-func (c *MemoryCache) Keys() []string {
+// 获取前缀匹配的key
+func (c *MemoryCache) Keys(prefix string) []string {
 	var arr = make([]string, 0)
-	var now = utils.Timestamp()
-	for _, bucket := range c.storage.Buckets {
-		bucket.RLock()
-		for k, v := range bucket.Map {
-			if c.valid(now, v.ExpireAt) {
-				arr = append(arr, k)
+	var now = time.Now().UnixMilli()
+	for _, b := range c.storage {
+		b.Lock()
+		for _, v := range b.Heap.Data {
+			if !v.Expired(now) && (prefix == "*" || strings.HasPrefix(v.Key, prefix)) {
+				arr = append(arr, v.Key)
 			}
 		}
-		bucket.RUnlock()
+		b.Unlock()
 	}
 	return arr
 }
@@ -110,15 +151,38 @@ func (c *MemoryCache) Keys() []string {
 // 获取有效元素个数
 func (c *MemoryCache) Len() int {
 	var num = 0
-	var now = utils.Timestamp()
-	for _, bucket := range c.storage.Buckets {
-		bucket.RLock()
-		for _, v := range bucket.Map {
-			if c.valid(now, v.ExpireAt) {
+	var now = time.Now().UnixMilli()
+	for _, b := range c.storage {
+		b.Lock()
+		for _, v := range b.Heap.Data {
+			if !v.Expired(now) {
 				num++
 			}
 		}
-		bucket.RUnlock()
+		b.Unlock()
 	}
 	return num
+}
+
+type bucket struct {
+	sync.Mutex
+	Map  map[string]*types.Element
+	Heap *heap.Heap
+}
+
+// 过期时间检查
+func (c *bucket) expireTimeCheck(maxNum int, maxCap int) {
+	c.Lock()
+	defer c.Unlock()
+
+	var now = time.Now().UnixMilli()
+	var num = 0
+	for c.Heap.Len() > 0 && c.Heap.Front().Expired(now) && num < maxNum {
+		delete(c.Map, c.Heap.Pop().Key)
+		num++
+	}
+	for c.Heap.Len() > maxNum && num < maxNum {
+		delete(c.Map, c.Heap.Pop().Key)
+		num++
+	}
 }
