@@ -3,65 +3,47 @@ package benchmark
 import (
 	"github.com/lxzan/memorycache"
 	"github.com/lxzan/memorycache/internal/utils"
-	"sync"
 	"testing"
 	"time"
 )
 
-const benchcount = 10000
+const benchcount = 1000000
 
 var benchkeys = make([]string, 0, benchcount)
 
 func init() {
 	for i := 0; i < benchcount; i++ {
-		benchkeys = append(benchkeys, utils.Alphabet.Generate(16))
+		benchkeys = append(benchkeys, string(utils.AlphabetNumeric.Generate(16)))
 	}
 }
 
 func BenchmarkSet(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		var mc = memorycache.New(memorycache.WithSegment(16))
-		var wg = sync.WaitGroup{}
-		wg.Add(2)
-		go func() {
-			var d = utils.Rand.Intn(5)
-			for j := 0; j < benchcount/2; j++ {
-				mc.Set(benchkeys[j], 1, time.Duration(d)*time.Second)
-			}
-			wg.Done()
-		}()
-		go func() {
-			var d = utils.Rand.Intn(5)
-			for j := benchcount / 2; j < benchcount; j++ {
-				mc.Set(benchkeys[j], 1, time.Duration(d)*time.Second)
-			}
-			wg.Done()
-		}()
-		wg.Wait()
+	var f = func(n, count int) {
+		var mc = memorycache.New(memorycache.WithBucketNum(16))
+		for i := 0; i < n; i++ {
+			var key = benchkeys[i%count]
+			mc.Set(key, 1, time.Hour)
+		}
 	}
+
+	b.Run("10000", func(b *testing.B) { f(b.N, 10000) })
+	b.Run("1000000", func(b *testing.B) { f(b.N, 1000000) })
 }
 
 func BenchmarkGet(b *testing.B) {
-	var mc = memorycache.New(memorycache.WithSegment(16))
-	for j := 0; j < benchcount; j++ {
-		mc.Set(benchkeys[j], 1, -1)
+	var f = func(n, count int) {
+		var mc = memorycache.New(memorycache.WithBucketNum(16))
+		for i := 0; i < count; i++ {
+			var key = benchkeys[i]
+			mc.Set(key, 1, time.Hour)
+		}
+
+		for i := 0; i < n; i++ {
+			var key = benchkeys[i%count]
+			mc.Get(key)
+		}
 	}
 
-	for i := 0; i < b.N; i++ {
-		var wg = sync.WaitGroup{}
-		wg.Add(2)
-		go func() {
-			for j := 0; j < benchcount/2; j++ {
-				mc.Get(benchkeys[j])
-			}
-			wg.Done()
-		}()
-		go func() {
-			for j := benchcount / 2; j < benchcount; j++ {
-				mc.Get(benchkeys[j])
-			}
-			wg.Done()
-		}()
-		wg.Wait()
-	}
+	b.Run("10000", func(b *testing.B) { f(b.N, 10000) })
+	b.Run("1000000", func(b *testing.B) { f(b.N, 1000000) })
 }
